Initial Commit 🎉
rustube is a YouTube downloader written in Rust. It's based on pytube, but has a slightly different public API and internally makes heavy use of 'rusty' features. Rusts type safety in combination with the tokio runtime make this youtube downloader blazing fast.

Currently rustube is still WIP! But the public API will probably not see huge changes in the future.

The goal of rustube is to become a highly customizable YouTube downloader implementation, which also provides bindings to other languages like python.

- [x] successfully download a video
- [x] provide an async API
- [ ] provide a blocking API
- [ ] provide good error handling
- [ ] become highly customizable through features
- [ ] provide FFI bindings
- [ ] have a robust test suite