Fixed a problem regarding YouTube's new consent mechanism
YouTube recently changed the way some users have to agree to cookies, so that some users (probably only european users) are redirected to a dedicated site, that asks for cookie approval. That lead to an error, since the watch html didn't contain the correct information anymore.
This error can be fixed by supplying some headers and cookies.

Closes #8